Rework namespacedinformer into filtering by project name
- Rename to projectinformer
- Filter by label that specifies project name in Project CRD and Objects
  Labels, instead of namespace
