+++ pkg/multiproject/projectinformer/projectinformer_test.go
@@ -1,4 +1,4 @@
-package namespacedinformer
+package projectinformer
 
 import (
 	"testing"
@@ -6,17 +6,20 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/tools/cache"
+	"k8s.io/ingress-gce/pkg/flags"
 )
 
-// TestNamespacedInformer_AddEventHandler verifies that the
-// namespacedinformer.AddEventHandler method does not return an error.
-func TestNamespacedInformer_AddEventHandler(t *testing.T) {
+// TestProjectInformer_AddEventHandler verifies that the
+// projectinformer.AddEventHandler method does not return an error.
+func TestProjectInformer_AddEventHandler(t *testing.T) {
+	flags.F.MultiProjectCRDProjectNameLabel = "project-name-label"
+
 	sharedInformer := cache.NewSharedIndexInformer(nil, &corev1.Pod{}, 0, nil)
-	namespacedInformer := NewNamespacedInformer(sharedInformer, "test-namespace")
+	projectInformer := NewProjectInformer(sharedInformer, "test-project")
 
 	handler := cache.ResourceEventHandlerFuncs{}
 
-	_, err := namespacedInformer.AddEventHandler(handler)
+	_, err := projectInformer.AddEventHandler(handler)
 	if err != nil {
 		t.Fatalf("Failed to add event handler: %v", err)
 	}
@@ -25,20 +28,22 @@ func TestNamespacedInformer_AddEventHandler(t *testing.T) {
 // TestNamespacedInformer_AddEventHandlerWithResyncPeriod verifies that the
 // namespacedinformer.AddEventHandlerWithResyncPeriod method does not return an
 // error.
-func TestNamespacedInformer_AddEventHandlerWithResyncPeriod(t *testing.T) {
+func TestProjectInformer_AddEventHandlerWithResyncPeriod(t *testing.T) {
+	flags.F.MultiProjectCRDProjectNameLabel = "project-name-label"
+
 	testCases := []struct {
 		desc         string
-		namespace    string
+		projectName  string
 		resyncPeriod time.Duration
 	}{
 		{
 			desc:         "Add event handler with resync period",
-			namespace:    "test-namespace",
+			projectName:  "test-project",
 			resyncPeriod: time.Minute,
 		},
 		{
 			desc:         "Add event handler with zero resync period",
-			namespace:    "test-namespace",
+			projectName:  "test-project",
 			resyncPeriod: 0,
 		},
 	}
@@ -49,10 +54,10 @@ func TestNamespacedInformer_AddEventHandlerWithResyncPeriod(t *testing.T) {
 			t.Parallel()
 
 			sharedInformer := cache.NewSharedIndexInformer(nil, &corev1.Pod{}, 0, nil)
-			namespacedInformer := NewNamespacedInformer(sharedInformer, tc.namespace)
+			projectInformer := NewProjectInformer(sharedInformer, tc.projectName)
 
 			handler := cache.ResourceEventHandlerFuncs{}
-			_, err := namespacedInformer.AddEventHandlerWithResyncPeriod(handler, tc.resyncPeriod)
+			_, err := projectInformer.AddEventHandlerWithResyncPeriod(handler, tc.resyncPeriod)
 			if err != nil {
 				t.Fatalf("Failed to add event handler with resync period: %v", err)
 			}
